requirements.txt
llama-cpp-python
sentence-transformers
faiss-cpu
numpy
pytesseract
pdf2image
gTTS
pygame
psutil